+++ mix.exs
@@ -33,7 +33,6 @@ defmodule ExAws.Mixfile do
       {:ex_doc, "~> 0.11.4", only: :dev},
       {:httpoison, "~> 0.8", optional: true},
       {:poison, "~> 1.2 or ~> 2.0", optional: true},
-      {:httpotion, "~> 2.0", optional: true},
       {:jsx, "~> 2.5", optional: true}
     ]
   end
